Implemented automatic generation of seat when a BUs is added to teh table

+++ src/main/java/com/example/busticketbooking/busticketbookingapi/dto/BusDto.java
@@ -1,13 +1,10 @@
 package com.example.busticketbooking.busticketbookingapi.dto;
 
 import com.example.busticketbooking.busticketbookingapi.entity.Route;
-import com.example.busticketbooking.busticketbookingapi.entity.Seat;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
 import lombok.Data;
 
-import java.util.List;
+import java.math.BigDecimal;
+
 
 @Data
 public class BusDto {
@@ -15,6 +12,6 @@ public class BusDto {
     private Integer capacity;
     private String type;
     private Route route;
-    private double fare;
+    private BigDecimal fare;
     private Long routeId;
 }
